Test Against .NET Framework 4.8
The point of having a "netstandard2.0" target is for compatibility with
.NET Framework, as there are no other supported "netstandard2.0" targets
currently in support. This means that we should be testing against
.NET Framework.

Previously the "netstandard2.0" target was tested using .NET Core 3.1,
however that runtime is no longer supported and is being removed by
default from the test runners.

Migrate the tests to work on .NET Framework 4.8 which allows me to
address both the issues above. Some ways that the "System.Net.Http"
library operated triggered some changes in how we store the intermediate
request data so it is available to validate test outcomes. Certain
handling around floating-point numbers seems to be different as well,
triggering some issues for some of the converters.

+++ src/Aydsko.iRacingData/Converters/StatusTimeStampConverter.cs
@@ -27,7 +27,7 @@ public class StatusTimeStampConverter : JsonConverter<DateTimeOffset>
     {
         if (value is DateTimeOffset instant)
         {
-            var rawValue = (instant - Epoch).TotalMilliseconds / 1000;
+            var rawValue = (decimal)(instant - Epoch).TotalMilliseconds / 1000;
             writer.WriteNumberValue(rawValue);
             return;
         }
